Connect to api
https://equinor-sds-si.atlassian.net/browse/GDHT-58

Fetching, running and submitting a test.
Submitting test is OK. BUT API needs to change image endpoint to be public. After that is done the app should show the image as well.
Disabled test for showing changelog, because it's not ready in back-end.
Refactored the timer to be in vars instead of in react-state. (TestPage)
Also added new error headers (red banner on top if some api-troubles)