feat: Handle unsent completed tests (#136)
Added functionality to store an unsent completed hearing test and let the user send it when the user is connected to the internet. A user will not be able to start a new hearing test if the user have an unsent hearing test on their device since a hearing test is based on previous hearing tests.

Closes #119