feat: Ignore `preDelayMs` if previous node was not heard
Currently, each node in the hearing test includes a preDelayMs value that specifies the amount of time to wait before playing the node's sound after the previous node has finished. However, if the patient did not hear the previous sound, there is no need to wait the full preDelayMs before playing the next sound.

This update will improve the efficiency of the hearing test by ignoring the preDelayMs value when the previous node was not heard, resulting in a shorter overall test time.

Closes #343